+++ Example/ReusableDemo/CollectionViewCells/MyColorSquareCell.swift
@@ -18,16 +18,20 @@ import Reusable
  * This view is NOT loaded from a NIB (but defined entierly by code),
  * that's why it's not annotated as `NibLoadable` but only `Reusable`
  */
-class MyColorSquareCell: UICollectionViewCell, Reusable {
-  private lazy var colorView: UIView = {
+final class MyColorSquareCell: UICollectionViewCell, Reusable {
+  fileprivate lazy var colorView: UIView = {
     let colorView = UIView()
-    colorView.autoresizingMask = [.FlexibleWidth, .FlexibleHeight]
+    #if swift(>=3.0)
+      colorView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
+    #else
+      colorView.autoresizingMask = [.FlexibleWidth, .FlexibleHeight]
+    #endif
     colorView.frame = self.contentView.bounds.insetBy(dx: 10, dy: 10)
     self.contentView.addSubview(colorView)
     return colorView
   }()
 
-  func fill(color: UIColor) {
+  func fill(_ color: UIColor) {
     self.colorView.backgroundColor = color
   }
 }