+++ Example/ReusableDemo/CollectionViewCells/MyStoryboardTextSquareCell.swift
@@ -18,9 +18,9 @@ import Reusable
  * `registerCell` is implemented, and as we won't need to call that method anyway
  * because the storyboard will register the cell for us, this makes no difference here.
  */
-class MyStoryboardTextSquareCell: UICollectionViewCell, Reusable {
-  @IBOutlet private weak var titleLabel: UILabel!
-  func fill(text: String) {
+final class MyStoryboardTextSquareCell: UICollectionViewCell, Reusable {
+  @IBOutlet fileprivate weak var titleLabel: UILabel!
+  func fill(_ text: String) {
     titleLabel.text = text
   }
 }