+++ Example/ReusableDemo/TableViewCells/MyStoryBoardIndexPathCell.swift
@@ -13,11 +13,11 @@ import Reusable
 // As we will never have to explicitly call tableView.register(…) to register it:
 // The Main.storyboard already auto-registers its cells without the need for additional code
 
-class MyStoryBoardIndexPathCell: UITableViewCell, Reusable {
-  @IBOutlet private weak var sectionLabel: UILabel!
-  @IBOutlet private weak var rowLabel: UILabel!
+final class MyStoryBoardIndexPathCell: UITableViewCell, Reusable {
+  @IBOutlet fileprivate weak var sectionLabel: UILabel!
+  @IBOutlet fileprivate weak var rowLabel: UILabel!
 
-  func fill(indexPath: NSIndexPath) {
+  func fill(_ indexPath: IndexPath) {
     sectionLabel.text = String(indexPath.section)
     rowLabel.text = String(indexPath.row)
   }