+++ Sources/View/NibLoadable.swift
@@ -30,7 +30,11 @@ public extension NibLoadable {
   /// By default, use the nib which have the same name as the name of the class,
   /// and located in the bundle of that class
   static var nib: UINib {
-    return UINib(nibName: String(self), bundle: Bundle(forClass: self))
+    #if swift(>=3.0)
+      return UINib(nibName: String(describing: self), bundle: Bundle(for: self))
+    #else
+      return UINib(nibName: String(self), bundle: NSBundle(forClass: self))
+    #endif
   }
 }
 
@@ -46,9 +50,15 @@ public extension NibLoadable where Self: UIView {
    - returns: A `NibLoadable`, `UIView` instance
    */
   static func loadFromNib() -> Self {
-    guard let view = nib.instantiate(withOwner: nil, options: nil).first as? Self else {
+    #if swift(>=3.0)
+      let view = nib.instantiate(withOwner: nil, options: nil).first
+    #else
+      let view = nib.instantiateWithOwner(nil, options: nil).first
+    #endif
+
+    guard let typedView = view as? Self else {
       fatalError("The nib \(nib) expected its root view to be of type \(self)")
     }
-    return view
+    return typedView
   }
 }